pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atlassampler"
version = "0.0.1"
authors = [
  { name="Chirag Modi", email="modichirag92@gmail.com" },
]
description = "Adapting trajectory length and stepsize for HMC in a delayed rejection framework."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
]


[project.optional-dependencies]
full = [
"jax",
"jaxlib",
"numpyro",
"optax",
"scipy"]

[project.urls]
Homepage = "https://github.com/modichirag/AtlasSampler/"

[tool.setuptools.packages.find]
include = ["src"]